modules/kuberay/kuberay-tests/pyproject.toml
[project]
requires-python = ">= 3.10"
name = "test"
version = "0.1"
dependencies = [
  "torch",
  "torchvision",
  "ray[default]==2.46.0",
  "pandas",
  "pyarrow",
  "python-dotenv",
  "cupy-cuda12x"
]

